refactor(Media content): Do just in time conversion to dataURIs
It is better to convert file URIs to data URIs only when we need to (i.e. while encoding to HTML).
It avoids unecessary compute and memory consumption.
